refactor(payroll): share ZIP progress hook and selectable table columns
